+++ src/modules/category.js
@@ -1,0 +1,17 @@
+const categoryList = () => {
+  const cats = [
+    "Other",
+    "Sport",
+    "Local",
+    "Politics",
+    "Economy",
+    "World",
+    "Entertainment",
+  ];
+
+  return cats.map((cat) => {
+    return { key: cat, text: cat, value: cat.toLowerCase() };
+  });
+};
+
+export default categoryList;
